recreate ad_create_update.ps1
A complete rework, Pulls information from the active directory to search for each contact on email. update it. if new contacts will wait till the end and prompt for what client they are to be added to.
creates an exclusion list for any that you mark do not create and won't prompt for client allocation next time it runs.